Fix download api error and use mapster

+++ src/Seventh.Resource.Services/AssetDownloadService.cs
@@ -1,10 +1,11 @@
 ﻿using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Seventh.Resource.Common.Classes.Options;
 using Seventh.Resource.Common.Crypts;
+using Seventh.Resource.Common.Entities;
 using Seventh.Resource.Common.Extensions;
 using Seventh.Resource.Common.Helpers;
+using Seventh.Resource.Common.Options;
 
 namespace Seventh.Resource.Services
 {
@@ -12,16 +13,17 @@ namespace Seventh.Resource.Services
     {
         private readonly AssetSortService _sortService;
         private readonly HttpClient _client;
-        private readonly AssetPath _assetPath;
+        private readonly PathOption _pathOption;
 
-        public AssetDownloadService(AssetDownloadClient client, AssetSortService sortService,ResourceLocation option)
+        public AssetDownloadService(AssetDownloadClient client, AssetSortService sortService, ResourceLocation option)
         {
             _sortService = sortService;
             _client = client.Client;
-            _assetPath = option.PathOption.AssetPath;
+
+            _pathOption = option.PathOption;
         }
 
-        public async Task<(bool result, string savePath, string sortedSavePath)>
+        public async Task<(bool result, AssetFileInfo info)>
             TryDownloadAtRevisionAndSortAsync(string fileName, int revision, bool needHash)
         {
             if (needHash)
@@ -29,17 +31,20 @@ namespace Seventh.Resource.Services
                 fileName = FileNameConverter.ToWithHashName(fileName);
             }
 
-            var (result, savePath) = await TryDownloadAtRevisionAsync(fileName, revision,false);
+            var (result, savePath) = await TryDownloadAtRevisionAsync(fileName, revision, needHash);
             if (!result)
             {
-                return (false, null, null);
+                return (false, null);
             }
 
-            return await DecryptAndSort(fileName, savePath);
+            var info = await DecryptAndSort(fileName, string.Concat(_pathOption.RootPath, savePath));
+            info.Revision = revision;
+            info.MirrorSavePath = info.MirrorSavePath.Replace(_pathOption.RootPath, string.Empty);
+            info.SortedSavePath = info.SortedSavePath.Replace(_pathOption.RootPath, string.Empty);
+            return (true, info);
         }
 
-
-        public async Task<(bool result, string savePath, string sortedSavePath)>
+        public async Task<(bool result, AssetFileInfo info)>
             TryDownloadAtMirrorAndSortAsync(string fileName, bool needHash)
         {
             if (needHash)
@@ -47,35 +52,40 @@ namespace Seventh.Resource.Services
                 fileName = FileNameConverter.ToWithHashName(fileName);
             }
 
-            var (result,savePath) = await TryDownloadAtMirrorAsync(fileName, true);
+            var (result, savePath) = await TryDownloadAtMirrorAsync(fileName, needHash);
             if (!result)
             {
-                return (false,null,null);
+                return (false, null);
             }
-
-            return await DecryptAndSort(fileName, savePath);
+            
+            var info = await DecryptAndSort(fileName, string.Concat(_pathOption.RootPath, savePath));
+            info.Revision = 0;
+            info.MirrorSavePath = info.MirrorSavePath.Replace(_pathOption.RootPath, string.Empty);
+            info.SortedSavePath = info.SortedSavePath.Replace(_pathOption.RootPath, string.Empty);
+            return (true,info);
         }
 
         public async Task<(bool result, string savePath)>
-            TryDownloadAtRevisionAsync(string fileName, int revision ,bool needHash)
+            TryDownloadAtRevisionAsync(string fileName, int revision, bool needHash)
         {
             if (needHash)
             {
                 fileName = FileNameConverter.ToWithHashName(fileName);
             }
 
-            var savePath = _assetPath.RevMirrorAssetPath
+            var savePath = _pathOption.AssetPath.RevMirrorAssetPath
                 .AppendAndCreatePath(revision.ToString()).AppendPath(fileName);
 
             if (File.Exists(savePath))
             {
-                return (true, savePath);
+                return (true, savePath.Replace(_pathOption.RootPath,string.Empty));
             }
 
             var response = await _client.GetAsync(fileName);
-            return !response.IsSuccessStatusCode 
-                ? (false, null) 
-                : (true, await SaveFile(fileName, savePath, response));
+            return !response.IsSuccessStatusCode
+                ? (false, null)
+                : (true, (await SaveFile(fileName, savePath, response))
+                    .Replace(_pathOption.RootPath,string.Empty));
         }
 
         public async Task<(bool result, string savePath)>
@@ -86,7 +96,7 @@ namespace Seventh.Resource.Services
                 fileName = FileNameConverter.ToWithHashName(fileName);
             }
 
-            var savePath = _assetPath.GameMirrorAssetPath.AppendPath(fileName);
+            var savePath = _pathOption.AssetPath.GameMirrorAssetPath.AppendPath(fileName);
 
             if (File.Exists(savePath))
             {
@@ -94,14 +104,15 @@ namespace Seventh.Resource.Services
             }
 
             var response = await _client.GetAsync(fileName);
-            return !response.IsSuccessStatusCode 
-                ? (false, null) 
-                : (true, await SaveFile(fileName, savePath, response));
+            return !response.IsSuccessStatusCode
+                ? (false, null)
+                : (true, (await SaveFile(fileName, savePath, response))
+                    .Replace(_pathOption.RootPath,string.Empty));
         }
 
-        private async Task<string> SaveFile(string fileName, string savePath,HttpResponseMessage response)
+        private async Task<string> SaveFile(string fileName, string savePath, HttpResponseMessage response)
         {
-            var tempSavePath = _assetPath.DownloadTempRootPath.AppendPath(fileName);
+            var tempSavePath = _pathOption.AssetPath.DownloadTempRootPath.AppendPath(fileName);
             await using var fileStream = File.OpenWrite(tempSavePath);
             await response.Content.CopyToAsync(fileStream);
             fileStream.Close();
@@ -111,21 +122,53 @@ namespace Seventh.Resource.Services
             return savePath;
         }
 
-        private async Task<(bool result, string savePath, string sortedSavePath)>
+        private async Task<AssetFileInfo>
             DecryptAndSort(string fileName, string savePath)
         {
             var encVersion = AssetCrypt.IdentifyEncVersion(fileName);
+            var encFileName = fileName;
             fileName = AssetCryptHelper.Rename(fileName, encVersion);
             var sortedSavePath = await _sortService.SortAsync(fileName);
+
+            if (File.Exists(sortedSavePath))
+            {
+                return new AssetFileInfo
+                {
+                    FileName = encFileName,
+                    RealFileName = fileName,
+                    FileSize = new FileInfo(savePath).Length,
+                    RealFileSize = new FileInfo(sortedSavePath).Length,
+                    MirrorSavePath = savePath,
+                    SortedSavePath = sortedSavePath
+                };
+            }
+
             if (encVersion == AssetCrypt.EncVersion.NoEnc)
             {
                 File.Copy(savePath, sortedSavePath);
-                return (true, savePath, sortedSavePath);
+                return new AssetFileInfo
+                {
+                    FileName = encFileName,
+                    RealFileName = fileName,
+                    FileSize = new FileInfo(savePath).Length,
+                    RealFileSize = new FileInfo(sortedSavePath).Length,
+                    MirrorSavePath = savePath,
+                    SortedSavePath = sortedSavePath
+                };
             }
 
             await AssetCryptHelper.DecryptAsync(savePath, sortedSavePath, encVersion,
                 AssetCryptHelper.IdentifyShouldLz4(fileName));
-            return (true, savePath, sortedSavePath);
+
+            return new AssetFileInfo
+            {
+                FileName = encFileName,
+                RealFileName = fileName,
+                FileSize = new FileInfo(savePath).Length,
+                RealFileSize = new FileInfo(sortedSavePath).Length,
+                MirrorSavePath = savePath,
+                SortedSavePath = sortedSavePath
+            };
         }
 
         //public async Task DownloadCard(int cardId, FileSizeVersion sizeVersion = FileSizeVersion.Large)
